remove styles foleder
